Created Interface and Stage modules
